requirements.txt
configargparse==1.2.3
h5py==3.1.0
keras==2.4.3
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.2
scikit-image==0.17.2
scikit-learn==0.24
tqdm==4.49.0
tensorflow-gpu==2.5.0
ml_collections
autologging